updating kodi19 fix to pass automated tests
