+++ resources/language/resource.language.nl_nl/strings.po
